app/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
yoyo-migrations==8.2.0